Update ekko-lightbox.js
I was having an issue where the class "modal-open" on the body was not being removed once I exited the lightbox therefore was hiding my scroll bar and to retrieve it, I had to continuously refresh the page. This solves that issue. If there's a better way to solve this problem, feel free to do it another way.

Also, 'max-width' works better when the browser width is small such as on a mobile. I was getting some extra padding, after making the change, everything works fine.